Carrito Terminado

+++ JewilryGenNHibernate/CEN/JoyeriaJewirly/LineaPedidoCEN.cs
@@ -39,7 +39,7 @@ public ILineaPedidoCAD get_ILineaPedidoCAD ()
         return this._ILineaPedidoCAD;
 }
 
-public void ModificarLinea (int p_LineaPedido_OID, int p_unidades, int p_precio)
+public void ModificarLinea (int p_LineaPedido_OID, int p_unidades, float p_precio)
 {
         LineaPedidoEN lineaPedidoEN = null;
